added the vmath module to replace umath
New vmath module for simple math functions which support both uncertainties and vectorisation; basically the best bits of numpy and uncertainties.umath.

Updated stellar and orbital to use this. Much needed for stellar.black_body_spectral_radiance() as this will be used with an array of lambdas whenever it's called (**a lot** when generating the ebop_maven synth-test-ds)