refactored Task to better support new Tasks (relates to #2)
Reworked Task base class so that it just publishes a single run() function which does the basic job of creating an input file and synchronously running jktebop then optionally reading the output file and cleaning down any files. This should work well for up coming Tasks 3, 8 & 9 and also works for the existing Task2 with a simple override to handle using a temp file. Updated the tests and task2 template.